+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 v1.4.21
 ---------------
 - Add Character selection for Magic Items in Campaign
-- Add Magic Items#index
+- Add Magic Items#index and show pages
 - Add Magic Item deletion
 - Add Magic Item property Not Include in Count
 - Allow multiple DM to join Campaigns